#20 split coverage attribution responsibilities
