+++ packages/core/src/istanbul.ts
@@ -2,35 +2,14 @@ import { readFile } from "node:fs/promises";
 import path from "node:path";
 
 import { normalizePathForMatch } from "./utils";
-import type { CoverageMetric, CoverageUnknownReason, MethodDescriptor, SourceSpan } from "./types";
+import type { SourceSpan } from "./types";
+import type { BranchCoverageUnit, FileCoverage, FunctionCoverageUnit, StatementCoverageUnit } from "./coverageUnits";
 
 const MAX_COLUMN = Number.MAX_SAFE_INTEGER;
 
-interface StatementCoverageUnit {
-  span: SourceSpan;
-  hits: number;
-}
-
-interface BranchCoverageUnit {
-  span: SourceSpan;
-  hits: number[];
-}
-
-interface FunctionCoverageUnit {
-  span: SourceSpan;
-}
-
-export interface FileCoverage {
-  statements: StatementCoverageUnit[];
-  branches: BranchCoverageUnit[];
-  functions: FunctionCoverageUnit[];
-}
-
-export interface MethodCoverage {
-  coverage: CoverageMetric;
-  statementCoverage: CoverageMetric;
-  branchCoverage: CoverageMetric;
-}
+export { coverageForMethods } from "./coverageAttribution";
+export type { MethodCoverage } from "./coverageNormalization";
+export type { FileCoverage } from "./coverageUnits";
 
 type JsonRecord = Record<string, unknown>;
 
@@ -70,286 +49,6 @@ export async function parseCoverageReport(
   return records;
 }
 
-export function coverageForMethods(
-  methods: MethodDescriptor[],
-  fileCoverage: FileCoverage | undefined,
-  fileUnknownReason: CoverageUnknownReason = "file_unmatched"
-): MethodCoverage[] {
-  if (!fileCoverage) {
-    return methods.map((method) => unavailableMethodCoverage(method, fileUnknownReason));
-  }
-
-  const attributableMethods = buildAttributableMethods(methods, fileCoverage.functions);
-  const attributed = methods.map(() => ({
-    statements: [] as StatementCoverageUnit[],
-    branches: [] as BranchCoverageUnit[]
-  }));
-
-  for (const statement of fileCoverage.statements) {
-    const owner = findOwningMethodIndex(attributableMethods, statement.span);
-    if (owner !== null) {
-      attributed[owner].statements.push(statement);
-    }
-  }
-
-  for (const branch of fileCoverage.branches) {
-    const owner = findOwningMethodIndex(attributableMethods, branch.span);
-    if (owner !== null) {
-      attributed[owner].branches.push(branch);
-    }
-  }
-
-  return methods.map((method, index) =>
-    attributableMethods[index].fnMapConflict
-      ? unavailableMethodCoverage(method, "fnmap_conflict")
-      : computeMethodCoverage(method, attributed[index].statements, attributed[index].branches)
-  );
-}
-
-interface AttributableMethod {
-  span: SourceSpan;
-  fnMapConflict: boolean;
-}
-
-function buildAttributableMethods(
-  methods: MethodDescriptor[],
-  functions: FunctionCoverageUnit[]
-): AttributableMethod[] {
-  if (functions.length === 0) {
-    return methods.map((method) => ({
-      span: method.bodySpan,
-      fnMapConflict: false
-    }));
-  }
-
-  return methods.map((method) => {
-    const matchedFunction = matchFunctionCoverage(method.bodySpan, functions);
-    return {
-      span: matchedFunction?.span ?? method.bodySpan,
-      fnMapConflict: matchedFunction === null
-    };
-  });
-}
-
-function matchFunctionCoverage(
-  methodSpan: SourceSpan,
-  functions: FunctionCoverageUnit[]
-): FunctionCoverageUnit | null {
-  const exactMatches = functions.filter((entry) => spansEqual(entry.span, methodSpan));
-  if (exactMatches.length === 1) {
-    return exactMatches[0];
-  }
-  if (exactMatches.length > 1) {
-    return null;
-  }
-
-  const lineAlignedMatches = functions.filter((entry) => spansShareBoundaryLines(entry.span, methodSpan));
-  const overlappingLineAlignedMatches = lineAlignedMatches.filter((entry) => spansOverlap(entry.span, methodSpan));
-  if (overlappingLineAlignedMatches.length === 1) {
-    return overlappingLineAlignedMatches[0];
-  }
-  if (overlappingLineAlignedMatches.length > 1) {
-    return null;
-  }
-
-  const containingMatches = functions.filter((entry) =>
-    spanContains(entry.span, methodSpan) || spanContains(methodSpan, entry.span)
-  );
-  if (containingMatches.length === 1) {
-    return containingMatches[0];
-  }
-
-  return null;
-}
-
-function computeMethodCoverage(
-  method: MethodDescriptor,
-  statements: StatementCoverageUnit[],
-  branches: BranchCoverageUnit[]
-): MethodCoverage {
-  const statementCoverage = toStatementCoverageMetric(method, statements);
-  const branchCoverage = toBranchCoverageMetric(method, branches);
-
-  return {
-    coverage: combineCoverageMetrics(statementCoverage, branchCoverage),
-    statementCoverage,
-    branchCoverage
-  };
-}
-
-function unavailableMethodCoverage(method: MethodDescriptor, reason: CoverageUnknownReason): MethodCoverage {
-  return {
-    coverage: unknownCoverageMetric(reason),
-    statementCoverage: method.expectsStatementCoverage
-      ? unknownCoverageMetric(reason)
-      : structuralNaCoverageMetric(),
-    branchCoverage: method.expectsBranchCoverage
-      ? unknownCoverageMetric(reason)
-      : structuralNaCoverageMetric()
-  };
-}
-
-function unknownMethodCoverage(reason: CoverageUnknownReason): MethodCoverage {
-  return {
-    coverage: unknownCoverageMetric(reason),
-    statementCoverage: unknownCoverageMetric(reason),
-    branchCoverage: unknownCoverageMetric(reason)
-  };
-}
-
-function toStatementCoverageMetric(
-  method: MethodDescriptor,
-  statements: StatementCoverageUnit[]
-): CoverageMetric {
-  const percent = percentageOfCoveredStatements(statements);
-  if (percent !== null) {
-    return measuredCoverageMetric(percent);
-  }
-  return method.expectsStatementCoverage
-    ? unknownCoverageMetric("statement_unattributed")
-    : structuralNaCoverageMetric();
-}
-
-function toBranchCoverageMetric(
-  method: MethodDescriptor,
-  branches: BranchCoverageUnit[]
-): CoverageMetric {
-  const percent = percentageOfCoveredBranches(branches);
-  if (percent !== null) {
-    return measuredCoverageMetric(percent);
-  }
-  return method.expectsBranchCoverage
-    ? unknownCoverageMetric("branch_unattributed")
-    : structuralNaCoverageMetric();
-}
-
-function combineCoverageMetrics(statementCoverage: CoverageMetric, branchCoverage: CoverageMetric): CoverageMetric {
-  if (statementCoverage.status === "unknown") {
-    return unknownCoverageMetric(statementCoverage.unknownReason!);
-  }
-  if (branchCoverage.status === "unknown") {
-    return unknownCoverageMetric(branchCoverage.unknownReason!);
-  }
-  if (statementCoverage.status === "structural_na" && branchCoverage.status === "structural_na") {
-    return {
-      percent: 100,
-      status: "structural_na",
-      unknownReason: null
-    };
-  }
-  return measuredCoverageMetric(Math.min(statementCoverage.percent ?? 100, branchCoverage.percent ?? 100));
-}
-
-function measuredCoverageMetric(percent: number): CoverageMetric {
-  return {
-    percent,
-    status: "measured",
-    unknownReason: null
-  };
-}
-
-function structuralNaCoverageMetric(): CoverageMetric {
-  return {
-    percent: null,
-    status: "structural_na",
-    unknownReason: null
-  };
-}
-
-function unknownCoverageMetric(reason: CoverageUnknownReason): CoverageMetric {
-  return {
-    percent: null,
-    status: "unknown",
-    unknownReason: reason
-  };
-}
-
-function percentageOfCoveredStatements(statements: StatementCoverageUnit[]): number | null {
-  if (statements.length === 0) {
-    return null;
-  }
-
-  const covered = statements.filter((statement) => statement.hits > 0).length;
-  return (covered / statements.length) * 100;
-}
-
-function percentageOfCoveredBranches(branches: BranchCoverageUnit[]): number | null {
-  let totalBranches = 0;
-  let coveredBranches = 0;
-
-  for (const branch of branches) {
-    totalBranches += branch.hits.length;
-    coveredBranches += branch.hits.filter((hits) => hits > 0).length;
-  }
-
-  if (totalBranches === 0) {
-    return null;
-  }
-
-  return (coveredBranches / totalBranches) * 100;
-}
-
-function findOwningMethodIndex(methods: AttributableMethod[], span: SourceSpan): number | null {
-  let bestMatch: number | null = null;
-
-  for (let index = 0; index < methods.length; index += 1) {
-    const method = methods[index];
-    if (method.fnMapConflict) {
-      continue;
-    }
-    if (!spanContains(method.span, span) && !spanContainsPosition(method.span, span.startLine, span.startColumn)) {
-      continue;
-    }
-    if (bestMatch === null || spanContains(methods[bestMatch].span, method.span)) {
-      bestMatch = index;
-    }
-  }
-
-  return bestMatch;
-}
-
-function spanContains(container: SourceSpan, candidate: SourceSpan): boolean {
-  return comparePosition(
-    container.startLine,
-    container.startColumn,
-    candidate.startLine,
-    candidate.startColumn
-  ) <= 0 && comparePosition(candidate.endLine, candidate.endColumn, container.endLine, container.endColumn) <= 0;
-}
-
-function spanContainsPosition(span: SourceSpan, line: number, column: number): boolean {
-  return comparePosition(span.startLine, span.startColumn, line, column) <= 0 &&
-    comparePosition(line, column, span.endLine, span.endColumn) < 0;
-}
-
-function comparePosition(
-  leftLine: number,
-  leftColumn: number,
-  rightLine: number,
-  rightColumn: number
-): number {
-  if (leftLine !== rightLine) {
-    return leftLine - rightLine;
-  }
-  return leftColumn - rightColumn;
-}
-
-function spansEqual(left: SourceSpan, right: SourceSpan): boolean {
-  return left.startLine === right.startLine &&
-    left.startColumn === right.startColumn &&
-    left.endLine === right.endLine &&
-    left.endColumn === right.endColumn;
-}
-
-function spansShareBoundaryLines(left: SourceSpan, right: SourceSpan): boolean {
-  return left.startLine === right.startLine &&
-    left.endLine === right.endLine;
-}
-
-function spansOverlap(left: SourceSpan, right: SourceSpan): boolean {
-  return comparePosition(left.startLine, left.startColumn, right.endLine, right.endColumn) < 0 &&
-    comparePosition(right.startLine, right.startColumn, left.endLine, left.endColumn) < 0;
-}
 
 function parseStatements(statementMapValue: unknown, hitsValue: unknown): StatementCoverageUnit[] {
   if (!isRecord(statementMapValue) || !isRecord(hitsValue)) {
